requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Vision Models & Augmentation
timm>=0.9.0
opencv-python>=4.8.0
albumentations>=1.3.0
Pillow>=10.0.0
pycocotools>=2.0.7

# Hugging Face (NLP & datasets)
transformers>=4.30.0
datasets>=2.14.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
jiwer>=3.0.0

# NLP
nltk>=3.8.0
spacy>=3.7.0
sacrebleu>=2.3.0

# Data Processing & Visualization
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
statsmodels>=0.14.0
wordcloud>=1.9.0

# Web App & Demo
gradio>=4.0.0
streamlit>=1.28.0

# Experiment Tracking
wandb>=0.16.0
optuna>=3.4.0

# Jupyter & Notebooks
jupyter>=1.0.0
ipywidgets>=8.1.0
nbformat>=5.9.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0.0
